+++ packages/npm/less-plugins/packages/shared/src/config.ts
@@ -1,0 +1,20 @@
+import { findClosestFile } from './find-closest-file';
+import { LESS_PLUGINS_CONFIG_JS } from './constants';
+import merge from 'lodash.merge';
+
+export interface LessPluginsConfig {
+    hooks: {};
+}
+
+export const defaultConfig: LessPluginsConfig = { hooks: {} };
+
+export function loadConfig(dirPath: string): LessPluginsConfig {
+    let currentConfig = merge({}, defaultConfig);
+    try {
+        const configPath = findClosestFile({ dirPath, fileName: LESS_PLUGINS_CONFIG_JS });
+        if (configPath) {
+            currentConfig = merge(currentConfig, require(configPath));
+        }
+    } catch (e) {}
+    return currentConfig;
+}