+++ packages/npm/less-plugins/packages/shared/src/constants.ts
@@ -1,1 +1,2 @@
 export const FILE_NAME_PACKAGE_JSON = 'package.json';
+export const LESS_PLUGINS_CONFIG_JS = 'less.plugins.config.js';