+++ packages/npm/less-plugins/packages/css-logical/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.10
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.9
 
 - Update dependencies @less-plugins/shared@0.0.7