+++ packages/npm/less-plugins/packages/get-current-package-name/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.9
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.8
 
 - 修复没有 index.js 的报错问题