+++ packages/npm/less-plugins/packages/get-env/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.2
 
 - init get-env