+++ packages/npm/less-plugins/packages/get-hashed-version/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.9
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.8
 
 - 更多的函数内置进 shared