+++ packages/npm/less-plugins/packages/get-package-info/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.3
 
 - Update dependencies @less-plugins/shared@0.0.7