+++ packages/npm/less-plugins/packages/replace-properties/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.7
+
+- Update dependencies @less-plugins/shared@0.0.8
+
 ## 0.0.6
 
 - Update dependencies @less-plugins/shared@0.0.7